discord_simhash_bot_debug/requirements.txt
discord.py>=2.3.2
numpy>=1.26.0
sentence-transformers>=2.7.0
